refactor(udev): use generic device class instead of direct uaccess
See [1], generic device class is the recommended way to set permission
for debug probe. Unfortunately, the device class ID_DEBUG_APPLIANCE was
introduced in systemd v258. So I use a trick to be compatible with older
systemd: in older version, it will set uaccess for ENV{ID_DEBUG_APPLIANCE}=="hslink",
in systemd>=258, it do redundant operation, but since TAG+= is
idempotent, it's safe.

[1]: https://gitlab.archlinux.org/archlinux/packaging/packages/probe-rs/-/issues/4

Signed-off-by: Coelacanthus <uwu@coelacanthus.name>